+++ src/schemathesis/engine/run/unit/__init__.py
@@ -297,5 +297,7 @@ def get_strategy_kwargs(ctx: EngineContext, *, operation: APIOperation, phase: P
     phase_config = ctx.config.phases_for(operation=operation).get_by_name(name=phase.name)
     if isinstance(phase_config, FuzzingPhaseConfig) and phase_config.extra_data_sources.is_enabled:
         kwargs["extra_data_source"] = ctx.extra_data_source
+    elif isinstance(phase_config, ExamplesPhaseConfig) and ctx.extra_data_source is not None:
+        kwargs["extra_data_source"] = ctx.extra_data_source
 
     return kwargs